Interrupt HKD constraint resolution
If needs to solve a constraint, and can find exactly one matching
instance, it "auto-selects" the instance, and starts searching for the
instance's constraints. This is annoying if the instance is defined in
terms of some internal classes, all of which are built with generics.

Really, if a user wants @Eq (HKD User Maybe)@, that's the constraint
they should write; the internal implementation behind that should remain
invisible.